Adjust sanity checks for UMAP plots in "9.5 Combined analyses" section
A recent change in igraph::cluster_louvain() affects the clustering seen
on the 2nd and 3rd UMAP plots of the "9.5 Combined analyses" section.
The new version of the function not only permutes the labels of the 11
clusters seen on these plots but it also affects the content of each cluster
as well as the topology of the clustering itself. See issue #8 for the
details.

This commit adjusts the sanity checks performed behind the scene so that
they reflect the new topology and labels.